refactor: fix component names to match Svelte conventions
